应用前台新模板完成

+++ src/main/www/dao/BaseDao.java
@@ -22,7 +22,7 @@ public interface BaseDao<T> {
      * @return 影响的行数
      * @throws DataAccessException 数据库操作异常
      */
-    int insertObject(T obj) throws DataAccessException;
+    long insertObject(T obj) throws DataAccessException;
 
     /**
      * 用主键编号删除对象
